ae.utils.geometry: Fix sqrtx with long

+++ utils/geometry.d
@@ -22,7 +22,7 @@ enum TAU = 2*PI;
 
 auto sqrtx(T)(T x)
 {
-	static if (is(T : int))
+	static if (is(T : long))
 		return std.math.sqrt(cast(float)x);
 	else
 		return std.math.sqrt(x);
